aula_3-8

+++ app/Http/Controllers/Admin/PlanController.php
@@ -47,14 +47,25 @@ class PlanController extends Controller{
     }
     
     public function destroy($url){
-        $plan = $this->repository->where('url',$url)->first();
+        $plan = $this->repository
+            ->with('details')
+            ->where('url',$url)
+            ->first();
 
         if(!$plan){
             return redirect()->back();
-        }else{
-            $plan->delete();
-            return redirect()->route('plans.index');
         }
+
+        if ($plan->details->count() > 0) {
+            return redirect()
+                ->back()
+                ->with('error','Existe, detalhes vinculados a esse plano, portando delete os detalhes antes');
+                
+        }
+
+        $plan->delete();
+        return redirect()->route('plans.index');
+        
     }
 
     public function search(Request $request){
